requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
PyYAML>=6.0.0
ultralytics>=8.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
websockets>=12.0
